requirements.txt
fastapi
uvicorn
jinja2
python-multipart
reportlab
aiofiles
watchfiles
python-dotenv
